+++ CHANGELOG.md
@@ -1,6 +1,8 @@
 ## Unreleased
 
 + Link CoreFoundation on macOS
++ Tweak code style, use 2-spaces instead of tabs
++ New functions: `pushFont`
 
 ## v0.11.3